pyproject.toml
[project]
name = "hyper-track"
version = "0.1.0"
description = "Hyperliquid perpetual futures wallet discovery and tracking engine"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
